Added tree generation. The trees themselves are for now still placeholders, but the algorithm for determining proper tree locations is already in place. The algorithm looks for spots that are not to steep, not too close to the racetrack or to other trees, and not under water.

+++ RobotRaceGradle/src/main/java/terrain/HeightMap.java
@@ -20,5 +20,5 @@ public interface HeightMap {
      * @param y The y-coordinate in meters.
      * @return The z-coordinate of the terrain at (x, y) in meters.
      */
-    float heightAt(float x, float y);
+    float heightAt(double x, double y);
 }
